+++ CHANGELOG.md
@@ -1,12 +1,19 @@
 # LogbookWrapped Changelog
 
 *Last updated: 2026-04-17* 
-*Current Version: v0.8.2*
+*Current Version: v0.8.3*
 
 ## Overview
 This document tracks past changes starting with v0.8.1.
 
 ## Changelog
+
+### Release v0.8.3
+- Screenshot Updates
+- Swap pages 2 and 3
+- Export page ui uppdates
+- Refine support for ForeFlight and Garmin 
+- Add support for LogTen and MyFlightbook
 
 ### Release v0.8.2
 - Vertically center items in cell on Page 8